Fix findCellForSpan to work with ADW's VacantCells
When long-pressing to select a new desktop item to add, the item should be added
at or near the position the user long-pressed. This worked on the stock launcher
because the way the VacantCells was built up centered around the current cell.

ADW uses a different method of building the VacantCells because the stock method
is really confusing and ANRs/infinite loops on larger grids.

This change to findCellForSpan does the find-nearest work so desktop items are
once again placed at or near where the user long-pressed.